+++ Forge/src/main/java/de/markusbordihn/easymobfarm/block/entity/MobFarmBlockEntityWrapper.java
@@ -26,6 +26,9 @@ import net.minecraft.core.Direction;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraftforge.common.capabilities.Capability;
+import net.minecraftforge.common.capabilities.ForgeCapabilities;
+import net.minecraftforge.common.util.LazyOptional;
 import net.minecraftforge.items.IItemHandler;
 import net.minecraftforge.items.wrapper.SidedInvWrapper;
 
@@ -43,5 +46,15 @@ public class MobFarmBlockEntityWrapper extends MobFarmBlockEntity {
   @Override
   protected IItemHandler createUnSidedHandler() {
     return new SidedInvWrapper(this, Direction.DOWN);
+  }
+
+  @Override
+  public <T> LazyOptional<T> getCapability(Capability<T> capability, Direction direction) {
+    if (capability == ForgeCapabilities.ITEM_HANDLER) {
+      if (direction != null && direction != Direction.UP) {
+        return LazyOptional.of(() -> new SidedInvWrapper(this, direction)).cast();
+      }
+    }
+    return super.getCapability(capability, direction);
   }
 }